gcp/viet-soccer/youtube-comment/requirements.txt
Flask==2.0.0
gunicorn==20.1.0
grpcio==1.37.1
protobuf==3.17.0
google-auth-oauthlib==0.4.4
google-auth==1.30.1
google-api-python-client==2.5.0
google-cloud-pubsub==2.5.0

